panel: Add 4 random sprites at end of green panel file to inflate size
The `pop2-1` file is used by SW to allocate memory for the panel - meaning
it needs to be the largest of the files.  Bur our changes optimized some
of the sprites, making the file size decrease. The file might have also
been padded, not sure.

This patch restores the file status as the largest.